Move the parent_info field from Window to BrowsingContext.
